+++ app/src/main/java/com/rest/client/rest/events/RestResponseEvent.java
@@ -5,18 +5,18 @@ import android.support.annotation.Nullable;
 import io.realm.RealmObject;
 
 public final class RestResponseEvent {
-	private int         mId;
+	private long         mId;
 	private
 	@Nullable
 			RealmObject mDBItem;
 
-	public RestResponseEvent( int id, @Nullable RealmObject dbItem ) {
+	public RestResponseEvent( long id, @Nullable RealmObject dbItem ) {
 		mId = id;
 		mDBItem = dbItem;
 	}
 
 
-	public int getId() {
+	public long getId() {
 		return mId;
 	}
